+++ shopping-cart/src/main/java/digitalinnovation/one/experts/shoppingcart/models/Item.java
@@ -1,4 +1,4 @@
-package digitalinnovation.one.experts.shoppingcart.model;
+package digitalinnovation.one.experts.shoppingcart.models;
 
 import org.springframework.data.redis.core.RedisHash;
